fix(docs, concerns): update documentation, rename concern
fix broken link in documentation
rename SaveHelper to DefaultActions
completely abstact the index and update controller actions

[#119133327]
