+++ public/index.html
@@ -439,7 +439,7 @@
 <li>Method: PATCH</li>
 <li>Authentication: required</li>
 </ul>
-<h2 id="all-transactions">Pagination&nbsp;</h2>
+<h2 id="pagination">Pagination&nbsp;</h2>
 <blockquote>
 <p>To paginate customers or transactions, use this:</p>
 </blockquote>
@@ -470,7 +470,7 @@
 <li>For page number pass in the parameter <strong>page.&nbsp;</strong>If no page nmber is specified, the first page is returned by default.</li>
 <li>For number of records per pag, pass in the parameter <strong>limit.&nbsp;</strong>If nothing is passed in, 20 is used as the default value.</li>
 </ul>
-<h2 id="all-transactions">Search&nbsp;</h2>
+<h2 id="search">Search&nbsp;</h2>
 <blockquote>
 <p>To search&nbsp;customers:</p>
 </blockquote>
